responses to allen's comments
